+++ application/admin/view/article/index.html
@@ -126,7 +126,7 @@
                 layer.open({
                     type: 2
                     ,title: '添加文章'
-                    ,content: '/article/add'
+                    ,content: 'add'
                     ,maxmin: true
                     ,area: ['80%','60%']
                     ,btn: ['确定', '取消']
