+++ application/admin/view/index/index.html
@@ -75,8 +75,8 @@
         <!-- 侧边菜单 -->
         <div class="layui-side layui-side-menu">
             <div class="layui-side-scroll">
-                <div class="layui-logo" lay-href="/">
-                    <span>{$site.seo_title|default="草帽聚合支付"} · 后台管理</span>
+                <div class="layui-logo" lay-href="/admin.html">
+                    <span>{$site.seo_title|default="草帽聚合支付"}</span>
                 </div>
                 <!--/菜单列表-->
                 <ul class="layui-nav layui-nav-tree" lay-shrink="all" id="LAY-system-side-menu" lay-filter="layadmin-system-side-menu">
