+++ application/config.php
@@ -61,8 +61,30 @@ return [
     // +----------------------------------------------------------------------
 
     'cache'                  => [
-        // 驱动方式
-        'type'   => 'Redis'
+        // 使用复合缓存类型
+        'type'  =>  'complex',
+        // 默认使用的缓存
+        'default'   =>  [
+            // 驱动方式
+            'type'   => 'redis',
+            // 服务器地址
+            'host'       => '127.0.0.1',
+        ],
+        // 文件缓存
+        'file'   =>  [
+            // 驱动方式
+            'type'   => 'file',
+            // 设置不同的缓存保存目录
+            'path'   => CACHE_PATH,
+        ],
+        // redis缓存
+        'redis'   =>  [
+            // 驱动方式
+            'type'   => 'redis',
+            // 服务器地址
+            'host'       => '127.0.0.1',
+        ]
+
     ],
 
     // +----------------------------------------------------------------------
