+++ public/static/common/modules/common.js
@@ -3,12 +3,12 @@ layui.define(function(e) {
     var i = (layui.$, layui.layer, layui.laytpl, layui.setter, layui.view, layui.admin);
     i.events.logout = function() {
         i.req({
-            url: "login/logout",
+            url: "/admin/login/logout",
             type: "get",
             data: {},
             done: function(e) {
                 i.exit(function() {
-                    location.href = "login/index"
+                    location.href = "admin/login"
                 })
             }
         })
